requirements.txt
Flask==2.3.2
Pillow==10.0.1
python-dotenv==1.1.0
